Add authorised user pages after login
